requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Caching==2.1.0
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
openai==1.25.1
packaging==24.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
tqdm==4.66.2
typing_extensions==4.11.0
Werkzeug==3.0.2
